Attempt to fix signed sector variable

+++ fs/ext4/dev.c
@@ -43,7 +43,7 @@
 static block_dev_desc_t *ext4fs_block_dev_desc;
 static disk_partition_t part_info;
 
-int ext4fs_set_blk_dev(block_dev_desc_t *rbdd, int part)
+unsigned long ext4fs_set_blk_dev(block_dev_desc_t *rbdd, int part)
 {
 	ext4fs_block_dev_desc = rbdd;
 
@@ -60,7 +60,7 @@ int ext4fs_set_blk_dev(block_dev_desc_t *rbdd, int part)
 	return part_info.size;
 }
 
-int ext4fs_devread(int sector, int byte_offset, int byte_len, char *buf)
+int ext4fs_devread(unsigned long sector, int byte_offset, int byte_len, char *buf)
 {
 	char sec_buf[SECTOR_SIZE];
 	unsigned block_len;
@@ -69,7 +69,7 @@ int ext4fs_devread(int sector, int byte_offset, int byte_len, char *buf)
 	if ((sector < 0)
 	    || ((sector + ((byte_offset + byte_len - 1) >> SECTOR_BITS)) >=
 		part_info.size)) {
-		printf("%s read outside partition %d\n", __func__, sector);
+		printf("%s read outside partition %lu\n", __func__, sector);
 		return 0;
 	}
 
@@ -77,7 +77,7 @@ int ext4fs_devread(int sector, int byte_offset, int byte_len, char *buf)
 	sector += byte_offset >> SECTOR_BITS;
 	byte_offset &= SECTOR_SIZE - 1;
 
-	debug(" <%d, %d, %d>\n", sector, byte_offset, byte_len);
+	debug(" <%lu, %d, %d>\n", sector, byte_offset, byte_len);
 
 	if (ext4fs_block_dev_desc == NULL) {
 		printf("** Invalid Block Device Descriptor (NULL)\n");
